cmd/go: support overlays for synthesized packages.
The main missing piece here was supporting Stat in the overlay filesystem, in
the parts of the package code that determines whether an command line argument
is a file on disk or a directory.  so this change adds a Stat function to the
fsys package. It's implemented the same way as the already existing fsys.lstat
function, but instead of os.Lstat, it calls os.Stat on disk files.

Then, the change changes parts of the package code to use the overlay Stat
instead of the os package's Stat.

For #39958

Change-Id: I8e478ae386f05b48d7dd71bd7e47584f090623df