cmd/go: support cgo files in overlays
This requires rewriting the paths of the files passed to the C toolchain
to use the overlaid paths instead of the disk paths of files. Because
the directories of the overlaid paths don't exist in general, tools have
been updated to run in base.Cwd instead of the package directory.

For #39958

Change-Id: I8986de889f56ecc2e64fa69f5f6f29fa907408f9