cmd/compiler:  allowing inlining of functions with OCALLPART
OCALLPART is exported in its original form, which is as an OXDOT.

The body of the method value wrapper created in makepartialcall() was not being
typechecked, and that was causing a problem during escape analysis, so I added
code to typecheck the body.

The go executable got slightly bigger with this change (13598111 -> 13598905),
because of extra exported methods with OCALLPART (I believe), while the text size
got slightly smaller (9686964 -> 9686643).

This is mainly part of the work to make sure all function bodies can be exported
(for purposes of generics), but might as well fix the OCALLPART inlining bug as well.

Fixes #18493

Change-Id: If7aa055ff78ed7a6330c6a1e22f836ec567d04fd