cmd/go: disallow overlays of non-.go and non-go.mod files
It's difficult to guarantee that .c/.cpp/ etc files can be processed
properly in an overlay: Because .c files can include headers from
relative directories, we would have to recreate the file tree with
all relative files in the right place, and even then bugs might
pop up from unexpected places. So only allow overlays on Go and go.mod
files.

Assembly files should be fine too, but they can be added to the list
of allowed extensions later if we decide to suport them.

For #39958

Change-Id: I19ca027d50b91259175f22a5aa15e660b39661fe