[dev.typeparams] cmd/compile/internal/importer, types2: initial check-in of types2 and importer
This is a 1:1 copy of the importer and types2 (unreviewed) prototype version.
The plan is to check in this code wholesale (it runs and passes all tests) and
then review the code file-by-file via subsequent CLs.

Specifically, the next CLs are:

1) Add comment "// UNREVIEWED" to each file.
2) Review file-by-file and remove the above comment for each
   reviewed file.

We may intersperse reviewed bug fixes and other improvements. Once all
"// UNREVIEWED" comments are gone, the packages are fully reviewed.

Change-Id: I9bd0ad1a7e7188b501423483a44d18e623c0fe71