unsafe: add Slice and String headers
Using the reflect package for these is unsuitable for many users, and
the old Data member there is of the wrong type. So, in the standard
library, and in many other libraries elsewhere, including x/sys, the
same old struct definitions have been cut and pasted around, sloppily.
This is a problem because these are structures that we really want to be
intimately tied to the Go compiler and hence should live as part of the
library that the compiler ships. Unsafe already provides Pointer and
Sizeof and related functions and types for exactly this purpose; they're
intimately tied up with the compiler. So too should we have Slice and
String headers, for the same types of manual constructions that people
are doing already with unsafe.Pointer. So, this commit moves those
definitions from internal/unsafeheader to unsafe.

Fixes: #19367

Change-Id: I780cefe35167c123b7af51ccc680b7c6e8d5db6b