cmd/compile: interleave inlining and escape analysis in single bottom-up visit
Goal is to expose escape information (e.g., parameter does not escape,
parameter is transmitted to result) to permit better inlining heuristics
(favor inlining things that do escape parameters, in hopes of removing
the escape in a larger context).

The escape analysis test appears to regress,but this is an artifact of
not knowing that the closure itself goes dead because of inlining,
and reporting the escape (correctly) in code that goes dead.

The current compiler actually skips escape-analyzing the dead method,
but feeds it to SSA anyway for optimization and code generation,
then finally removes it during linker dead-code.

Change-Id: I7d680193b1b46ab9bbdb4c5077c35266f8349798