cmd/compile: remove dead closures after inlining
Inlining closures can eliminate all calls to the function,
but it remains in the list to be compiled, and is not eliminated
until the linker runs dead code.

This doesn't work for -dynlink; need to keep an eye on this for
when we start to use the same bottom-up visitor for scheduling
SSA compilation.

A quick benchmark (7 builds) for time saved showed essentially none
at all, so this CL may be pointless.

Change-Id: Iff8a26b33210c672bf7f412b0936e765e7abbcf1