os/signal: add -- before nohup subcommand
This removes any ambiguity about whether flags are being passed to
nohup or to the subcommand, since the C getopt_long function stops
parsing at a double hyphen. At least some flavors of nohup rely on the
"--" to be present to avoid parsing flags in the subcommand as its own
flags.

Change-Id: I7c8dbf1fd7671f0d902a402b15688d7d505aa563