reflect: set ptrToThis when asked for pointer to dynamic type
When a type is created by the program, and then PtrTo is used to ask
for a pointer to that type, record the newly created pointer type in
the ptrToThis field of the element type. This speeds up uses of PtrTo
in programs that use lots of dynamic types.

For #41993

Change-Id: Ife9b6694aa514f54258ebb09da70dddb0a1b361e