cmd/compile:  code for exporting and importing typeswitches
Exporting/importing typeswitch is tricky, because of the pseudo-variable
at the top of the switch (which is an ONONAME node) and per-case
variables of each different type, none of which are actually declared
with ODCL, but are interlinked with the same symbol for the
pseudo-variable. Also, the per-case variables have the same names, but
different vargens.

In order to distinguish the various cases, I have encoded an extra flag
after ONONAME during export. Also, I have encoded the per-case variables
as ONONAME with a different flag, rather than their original ONAME
(which wouldn't distinguish them). Let me know if you think I should use
a different approach (e.g. I could use a completely new Op value).

I have used an extra hash table in iimport to be able to find the
OTYPESW associated with the pseudo-variable and to find the definitions
of the per-case variables. I need to do this to correctly share the same
Sym among all the variables, and also set var.Name.Defn properly to the
OTYPESW node.

I'm not planning to write the code so OTYPESW can actually be inlined.
So, for now, I added a tempory test flag testExportTypeswitch that
allows functions with OTYPESW to be exported. If you run make.bash with
that flag set to true, you will see various such functions being
exported. If you increase inlineMaxBudget enough, one of these functions
will be imported, but then there will be a failure when inlining. I have
compared the AST trees at export and at import for these functions, and
they look identical (including all the relevant links between nodes).

To show that, I found the changes in fmt.go useful -- let me know if
those seem useful.

Change-Id: I0e1fff3cbc7b3b90ad67efcbc44583daa6d28031