cmd/compile: early devirtualization of interface method calls
DO NOT REVIEW
DO NOT SUBMIT

This CL makes two changes:

1. When inlining a function call "f()", if "f" contains exactly 1
"return" statement and doesn't name its result parameters, it's
inlined to declare+initialize the result value using the AST
representation that's compatible with staticValue.

2. After inlining, add a pass that looks for interface calls where we
can statically determine the interface value's concrete type. If such
a case is found, insert an explicit type assertion to the concrete
type so that escape analysis can see it.

Fixes #33160.

Change-Id: I36932c691693f0069e34384086d63133e249b06b