cmd/compile: fix recognition of unnamed return variables
In golang.org/cl/266199, I reused the existing code in inlining that
recognizes anonymous variables. However, it turns out that code
mistakenly recognizes anonymous and blank return parameters as named
when inlining a function from the same package.

The issue is funcargs (which is only used for functions parsed from
source) synthesizes ~r and ~b return parameter names, but funcargs2
(which is only used for functions imported from export data) does not.

This CL fixes the behavior so that anonymous and blank return
parameters are handled identically whether a function is inlined in
the same package or another package. It also adds a proper
cross-package test case demonstrating #33160 is fixed in both cases.

Change-Id: Iaa39a23f5666979a1f5ca6d09fc8c398e55b784c