cmd/compile: modify inlining heuristic for if; max arms, not sum
DO NOT SUBMIT

This bases inlining cost on non-loop path length, rather than total
number of nodes (at least for OIF).

This also accounts for exits and returns under if arms; those
terminate paths (i.e., if the max path ends in a panic/return,
then it only counts if it causes the inlining budget to be
exceeded; it is not the cost of the IF itself).

The reduction in cost, either use of max instead of sum or by
terminating a path with a panic/return, is capped at
inlineIfMaxBudget (68) to prevent extravagant growth in binary size.

Budgets tried: Infinity, 250, 125, 75, 68, 61, 0.
Less than the cost of  a call (57) would not obtain the desired
"call in each IF branch inlined" behavior.
More could be tried, but this is not bad.

Incorporates fix to accounting (from mistake in previous patch
sets) for if-constant case.

Inlining of reflect/flag.mustBeExportedSlow is suppressed
to ensure that it does get inlined into the fast path method
(which can make it too large to be inlined, itself).

------
For this CL:

Combined with the previous CL, measuring over
the "bent" benchmarks:

Text size increase:
median  0.5%
mean    0.7%
max     1.6%

Time increases (negative is faster) in two different benchmark runs
             first    second (rebooted)
Min         -11.3%     -11.8%
25%ile       -1.15%     -1.4%
Median       -0.4%      -0.45%
Mean         -0.8%      -1%
Max          +2.4%      +2.2%
>1%slower     1          3
<-2% faster   8         12
<-1% faster  23         26

Benchmarks appear in
https://docs.google.com/spreadsheets/d/11vcH1xYRuAR1X1J8mV4ZwG7dps73YRhcxX_rtsZKIkM/edit?usp=sharing
sheets "Tuesday PM times" and "Wed afternoon times" (far to the right).

This is an "update", not a "fix", because the best-for-all max-not-sum budget turns out to not
be adequate for more than 2 calls total, not the requested (in some comments) 3.

Updates #30548.

Change-Id: Ia91016ed9c4088917b3a3826b0a9223254392168