cmd/dist: make the logic for slow tests more consistent
Factor out the repeated “non-mobile, non-Plan- 9” predicate into a
variable, and use that variable in the several places where it seems
to apply today.

This results in the ../misc/reboot test being skipped on js/wasm,
which is fine (because nobody is bootstrapping the Go tool on WASM)
and may help mitigate slow / flaky TryBots (noticed on CL 267719).

Change-Id: If9322bd19e2dbdeb6ab2f08648e63f5620a09de0