crypto/tls: de-prioritize AES-GCM ciphers when lacking hardware support
Unless both the server and client would prefer to use an AES-GCM cipher
suite, we first try to select an AEAD cipher suite that is not AES-GCM
before falling back to negotiating the first preferred mutual cipher
suite as before. With the current implemented cipher suites, the only
considered cipher suite will be ChaCha20-Poly1305.

Client preference for AES-GCM is communicated by the first supported
cipher suite in the client hello being AES-GCM which implies the client
has hardware support.

We respect the configured cipher suite list for TLSv1.2 and earlier when
both PreferServerCipherSuites and CipherSuites are set on the Config.

This builds upon CL 262857 and includes its test suite.

Fixes #41181

Co-authored-by: Roland Shoemaker <roland@golang.org>
Change-Id: Id9cd7cc6e8e2530fe5bc89e5f1434ac0c4f5b43b